casting#	modified:   utils.py

+++ utils.py
@@ -166,7 +166,7 @@ def coin_loss(_alphas, _betas, epsilon):
                 # if (replace_alphas[i]):
                 # alpha = tf.stop_gradient(coin_coef_a(y_true[slc, :, :, i], y_pred[slc, :, :, i], epsilon))
                 U = float(alphas[i]) # val_mean + K.sum(flat_pred)
-                alpha = tf.stop_gradient(- 2 / U)
+                alpha = tf.cast(tf.stop_gradient(- 2 / U), tf.float64)
                 # else:
                 #     alpha = float(alphas[i])
 
@@ -174,7 +174,7 @@ def coin_loss(_alphas, _betas, epsilon):
                 # beta = tf.stop_gradient(coin_coef_b(y_true[slc, :, :, i], y_pred[slc, :, :, i], epsilon))
                 I = float(betas[i]) # K.sum(flat_true * flat_pred)
                 U = float(alphas[i]) # val_mean + K.sum(flat_pred)
-                beta = tf.stop_gradient(2 * I / U**2)
+                beta = tf.cast(tf.stop_gradient(2 * I / U**2), tf.float64)
                 # else:
                 #     beta = float(betas[i])
                 loss += K.sum((alpha * y_true[slc, :, :, i] * y_pred[slc, :, :, i]) +  (beta * y_pred[slc, :, :, i]))
